Metadata is Own File
Now, the metadata is its own file. This had to be done for the metadata
content is a proto now and thus metadata is accessed at different
places.
