+++ tensorboard/webapp/plugins/npmi/metadata.py
@@ -1,0 +1,48 @@
+# -*- coding: utf-8 -*-
+# Copyright 2020 The TensorFlow Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from tensorboard.compat.proto import summary_pb2
+
+from tensorboard.webapp.plugins.npmi import plugin_data_pb2
+
+PLUGIN_NAME = "npmi"
+
+
+def create_summary_metadata(description, title):
+    content = plugin_data_pb2.NpmiPluginData(title=title)
+    return summary_pb2.SummaryMetadata(
+        summary_description=description,
+        plugin_data=summary_pb2.SummaryMetadata.PluginData(
+            plugin_name=PLUGIN_NAME, content=content.SerializeToString(),
+        ),
+    )
+
+
+def parse_plugin_metadata(content):
+    """Parse summary metadata to a Python object.
+    Arguments:
+      content: The `content` field of a `SummaryMetadata` proto
+        corresponding to the scalar plugin.
+    Returns:
+      A `ScalarPluginData` protobuf object.
+    """
+    if not isinstance(content, bytes):
+        raise TypeError("Content type must be bytes")
+    result = plugin_data_pb2.NpmiPluginData.FromString(content)
+    return result
